+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/infrastructure/CorfuInterClusterReplicationServer.java
@@ -241,6 +241,7 @@ public class CorfuInterClusterReplicationServer implements Runnable {
 
         configureMetrics(opts, serverContext.getLocalEndpoint());
 
+        CorfuSaasEndpointProvider.init(serverContext.getPluginConfigFilePath());
         // Register shutdown handler
         Thread shutdownThread = new Thread(this::cleanShutdown);
         shutdownThread.setName("ShutdownThread");
